Code moved to examples/introspection/erc165
