+++ contracts/examples/introspection/erc165/IWhitelist.sol
