+++ contracts/test/examples/introspection/erc165/ERC20WhitelistedMock.sol
@@ -1,6 +1,6 @@
 pragma solidity ^0.5.10;
 import "@openzeppelin/contracts/ownership/Ownable.sol";
-import "./../../drafts/token/ERC20Whitelisted.sol";
+import "./../../../../examples/introspection/erc165/ERC20Whitelisted.sol";
 
 
 contract ERC20WhitelistedMock is ERC20Whitelisted, Ownable {
